feat: create `AssistantAndOrgListbox`
