+++ gui/src/components/AssistantAndOrgListbox/AssistantIcon.tsx
@@ -1,13 +1,13 @@
 import { ComputerDesktopIcon, SparklesIcon } from "@heroicons/react/24/outline";
 import { ProfileDescription } from "core/config/ProfileLifecycleManager";
-import { isLocalProfile } from "../../../util";
+import { isLocalProfile } from "../../util";
 
 export interface AssistantIconProps {
   assistant: ProfileDescription;
   size?: number;
 }
 
-export default function AssistantIcon({ assistant, size }: AssistantIconProps) {
+export function AssistantIcon({ assistant, size }: AssistantIconProps) {
   const sizeTw = size ?? 4;
   if (isLocalProfile(assistant)) {
     return (
