+++ gui/src/components/AssistantAndOrgListbox/ScopeSelect.tsx
@@ -4,12 +4,6 @@ import {
 } from "@heroicons/react/24/outline";
 import { UserCircleIcon } from "@heroicons/react/24/solid";
 import { useContext } from "react";
-import {
-  Listbox,
-  ListboxButton,
-  ListboxOption,
-  ListboxOptions,
-} from "../../components/ui/Listbox";
 import { useAuth } from "../../context/Auth";
 import { IdeMessengerContext } from "../../context/IdeMessenger";
 import { useAppDispatch, useAppSelector } from "../../redux/hooks";
@@ -17,6 +11,12 @@ import {
   selectCurrentOrg,
   setSelectedOrgId,
 } from "../../redux/slices/profilesSlice";
+import {
+  Listbox,
+  ListboxButton,
+  ListboxOption,
+  ListboxOptions,
+} from "../ui/Listbox";
 
 interface ScopeSelectProps {
   onSelect?: () => void;
@@ -54,15 +54,19 @@ export function ScopeSelect({ onSelect }: ScopeSelectProps) {
         <ListboxButton className="hover:bg-list-active hover:text-list-active-foreground w-full min-w-[140px] justify-between px-4 py-2 sm:min-w-[200px]">
           <div className="flex items-center gap-2">
             {selectedDisplay?.iconUrl ? (
-              <img src={selectedDisplay.iconUrl} alt="" className="h-5 w-5" />
+              <img
+                src={selectedDisplay.iconUrl}
+                alt=""
+                className="h-3.5 w-3.5 rounded-full"
+              />
             ) : (
-              <CurScopeEntityFallBackIcon className="h-5 w-5" />
+              <CurScopeEntityFallBackIcon className="h-3.5 w-3.5" />
             )}
             <span className="truncate">
               {selectedDisplay?.name || "Select Organization"}
             </span>
           </div>
-          <ChevronUpDownIcon className="h-5 w-5" aria-hidden="true" />
+          <ChevronUpDownIcon className="h-3.5 w-3.5" aria-hidden="true" />
         </ListboxButton>
 
         <ListboxOptions className="z-[1000] min-w-[140px] pt-0.5 sm:min-w-[200px]">
@@ -70,9 +74,13 @@ export function ScopeSelect({ onSelect }: ScopeSelectProps) {
             <ListboxOption key={org.id} value={org.id} className="py-2">
               <div className="flex items-center gap-2">
                 {org.iconUrl ? (
-                  <img src={org.iconUrl} alt="" className="h-5 w-5" />
+                  <img
+                    src={org.iconUrl}
+                    alt=""
+                    className="h-3.5 w-3.5 rounded-full"
+                  />
                 ) : (
-                  <BuildingOfficeIcon className="h-5 w-5" />
+                  <BuildingOfficeIcon className="h-3.5 w-3.5" />
                 )}
                 <span>{org.name}</span>
               </div>
