+++ gui/src/components/FreeTrialButton.tsx
@@ -1,11 +1,10 @@
 import { ChevronDownIcon } from "@heroicons/react/24/outline";
 import { FreeTrialStatus } from "core/control-plane/client";
 import { useContext } from "react";
-import { Button, SecondaryButton, vscButtonBackground } from "../..";
-import { IdeMessengerContext } from "../../../context/IdeMessenger";
-import { fontSize } from "../../../util";
-import { Listbox, ListboxButton, ListboxOptions, Transition } from "../../ui";
-import { useFontSize } from "../../ui/font";
+import { Button, SecondaryButton, vscButtonBackground } from ".";
+import { IdeMessengerContext } from "../context/IdeMessenger";
+import { fontSize } from "../util";
+import { Listbox, ListboxButton, ListboxOptions, Transition } from "./ui";
 
 interface ProgressBarProps {
   label: string;
@@ -78,8 +77,6 @@ interface FreeTrialButtonProps {
 export default function FreeTrialButton({
   freeTrialStatus,
 }: FreeTrialButtonProps) {
-  const smallFont = useFontSize(-3);
-  const tinyFont = useFontSize(-4);
   const ideMessenger = useContext(IdeMessengerContext);
 
   return (
